refactor: separate logic form displaying, change project structure
