+++ src/searchPage/cardsGroup.tsx
