+++ src/searchPage/utils/getFormattedPagesArray.ts
