+++ src/searchPage/utils/getPagesArray.ts
